Don't restart if DISPLAY is not set
In some cases, like when connecting from a remote terminal using
SSH, GTK can't be initialized and fails. In that case, this
daemon keeps being relaunched over and over.

This patch fixes it by checking first if it is possible to
initialize GTK, and only relaunching the daemon if there was
an error.

Fix https://github.com/snapcore/snapd-desktop-integration/issues/14