Always restart with a delay
This MR changes the way the daemon responds when it is killed:
it will always restart, but with a 5-second delay to avoid
consuming too many resources. This allows to guarantee that
it is always launched, no matter if the user begins with a
ssh session and, in the middle, changes to a graphical one.