DT-774: install required theme snap on notification click
When a gtk-theme snap is missing, snapd-desktop-integration
shows a notification allowing to install it. Unfortunately,
when that notification is moved to the list of "old"
notifications, it stops reacting.

The reason is that when a notification in that list is
clicked, the program that created it receives a 'default'
action, while the two actions defined in the notification
are 'yes' and 'no'.

This patch adds a 'default' action that also installs
the snap.

Fix https://github.com/snapcore/snapd-desktop-integration/issues/13