DT-880: fix "spam" in journalctl
By default, systemd manages the launch of snapd-desktop-integration.
Unfortunately, due to the design of snapd, it is launched as an
user-daemon with "wantedby=default.target", which means that it
is launched under any session, not only graphical ones. This
means that it will fail to run under text-only sessions (like
SSH or console sessions), but also in the GDM session (because
currently snapd doesn't support launching snaps for users with
a HOME folder outside /HOME).

Since the daemon is configured to relaunch in case of a problem,
this means that journalctl will be flooded with error messages
from the daemon launched in the GDM session (which will be
relaunched periodically after the failure).

Unfortunately, just setting the daemon to "don't relaunch" is
not an option, for two reasons:

* any bug in the daemon that makes it fail will stop it until
the user closes the session and enters again

* the desktop startup is "slow", and the first time that the
daemon is launched usually fails to connect to the window
manager.

This means that it is mandatory for the daemon to relaunch
on failure, if we want to ensure that it is launched during
desktop startup; but at the same time we need systemd to
avoid relaunching it on failure in the GDM session.

This MR solves this problem by setting the relaunch mode as
"on-success". This relaunches the daemon if it returns 0
on exit. This solves the first problem: stop relaunching
it on GDM.

To guarantee that the daemon is relaunched always in an
user session, it is not enough to add a "return 0;" at
the end of the code, because there are several other
cases where it can fail. A classic example is a memory access
error, that kills the program returning a non-zero value.

To protect the daemon from these cases, the first thing that
it does is a "fork();" call to create a child, and it is
this child the one that runs the daemon code itself. The parent,
in the meantime, just waits until the child dies, and in that
moment it returns zero. This way, it doesn't matter what was
the return value, systemd will always "see" a zero.

There are some corner cases that must be managed, like what
happens when systemd wants to kill the daemon (for example,
during shutdown, or when snapd updates the package). For
those cases, the parent process captures SIGTERM and SIGINT
signals, and if it receives any of them, it kills the child
and returns the correct value (128 + signal_id).