DT-906: avoid daemon being relaunched over and over
When the user enters a non-graphical session (like a SSH
session), the GTK initialization in the daemon fails, and it
exits with a "zero" error to ensure that systemd relaunches
it after two seconds. The result is that, in those cases,
the journal log is full of messages saying that snapd-desktop-
integration has failed and is being reloaded, one every two
seconds.

Also, since the systemd daemons are launched once per user,
instead of once per session, we must keep reloading it even
if the initialization fails, because it can happen that the
user, while having the SSH session opened, enters into a
desktop session. Not reloading it would mean that snapd-
desktop-integration wouldn't be loaded in that case.

Also, sometimes the daemon is launched "too quickly" by
systemd, and although it is a desktop session, the desktop
isn't ready and gkt_init_check() fails. In that case we
must force a reload to ensure that it is launched again
when the desktop is ready.

This patch fixes it by monitoring the logind DBus interface
to detect when a new session is opened.

It first calls gtk_init_check(). If it suceeds, then the
daemon is running inside a desktop session and it can work
as expected.

But if it fails, it can mean that we are inside a SSH session,
or in a desktop session but the daemon was launched "too
quickly". So in this case, the daemon tries to connect to the
login1 daemon, which manages the sessions, and checks all the
current sessions for the current user.

If any of the sessions is a graphical session (which is known
by its type being "x11", "wayland" or "mir"), then the daemon
knows that it must exit in order to be relaunched two seconds
after, when the desktop is fully ready.

Instead, if there is no graphical session opened for the current
user, the daemon concludes that it is running in a SSH, or in
a local, text mode TTY, so it will connect to the "session-new"
signal of the logind daemon and wait for new sessions being
opened. When that happens, it will check if that new session
is a desktop one. If it ins't (because, for example, the user
opened another SSH connection), the daemon will just continue
waiting for a new session. But if the new session is a desktop
one, then it will exit to ensure that systemd will relaunch it.

It is a must to exit and let systemd to relaunch it because
the environment variables are different when running in a
SSH/TTY session than in a desktop session.

Finally, if the daemon is unable to connect/read data from the
logind daemon, it will just exit, thus working like the current
code. This can happen if the snap isn't connected to the
"login-session-observe" interface.