Udeng-2978: put the right icon in the top left corner (#77)
* Put the right icon in the top-left corner

Due to the fact that snap doesn't set a .desktop file for the
snaps with the original name, but builds one with
snap-name_app-name.desktop, the desktop identification usually
fails. This makes GNotify to not be able to identify the .desktop
file and set the right icon in the top-left corner of the
notifications it sends.

This patch migrates the notification code to libnotify, which
allows to bypass this problem. But the original code is kept
inside an #ifdef to allow to revert to it in the future, when
this problem is fixed in snapd.