Fix cancelled refreshes
When a refresh was cancelled (for example, because the snap was
launched before it has completed the refresh), the snap is
reverted, but the current code fails to detect correctly this,
and snapd-desktop-integration could be left in an invalid state,
asking for already completed changes over and over again.

This patch fixes this bug.