Move the snapd-notices-monitor logic into its own class
This change moves the snapd notices monitor into its own class.
This object will keep a connection permanently open, even if
snapd closes it (for example, because it is updated), by opening
a new one. It also re-sends each notice event with exactly the
same format, thus allowing to connect to it with the same logic
than the original snapd notices monitor.