Refactored all the window progress code
This PR moves the GTK window and dock progress bars management
into their own classes, letting the `SdiRefreshMonitor` object
to process the events from snapd and just send signals with the
changes, signals that are connected to the new
`SdiProgressWindow` and `SdiProgressDock` objects, which are
responsible of creating a window if needed, update the progress
bars in each element, and close a window, for the former object,
and update the progress bar in the dock for the later.

This separation simplifies the code and, along with #97, allows
to isolate the event management code, which will allow to create
unitary tests for the process.