Fix theme change test
When launching the theme-change test, it is done through Weston
window manager, because it requires it. Unfortunately, if the
test fails and returns a non-zero value, Weston doesn't
propagate it, so it always returns a zero, thus cheating the
CI/CD pipeline and making it believe that the test passed.

This patch fixes this by checking the return value of the test
and sending a SIGKILL to Weston if the test failed; this makes
Weston to return a non-zero value, thus notifying to the CI/CD
pipeline that the test failed.