Add support for snapd abstract socket
Snapd has changed to an abstract socket when used from inside a
snap, so a change was required in snapd-glib. Now, the library
automagically detects if it is inside a snap and chooses, in
that case, the new abstract socket if it exists, or the old,
normal socket if the abstract doesn't exist, so the current code
to set the socket if we are inside a snap not only is not needed
anymore, but it will stop working in the next snapd version,
so it must be removed and let snapd-glib to choose the right
socket.