Don't show notifications with negative expiration time (#140)
* Don't show notifications with negative expiration time

Sometimes, snapd notifies us with an event that has a negative
time for a forced quit, when, in fact, there are no pending
updates for that snap. This is a very undesirable effect,
because it confuses the final user.

This patch ensures that no notification with a negative value
is shown.

fix https://github.com/canonical/snapd-desktop-integration/issues/135

Co-authored-by: Alessandro Astone <ales.astone@gmail.com>