Add a workaround for Ubuntu Budgie
Ubuntu Budgie has a bug in its notifications system: when the
user clicks on an action (for example, "Yes" for installing
new themes from the Snap Store), it sends the button action
(thus, "yes" action), but after that also sends the default
action. Instead, both Gnome and KDE only sends the button
action.

This results in odd behavior in this system: since the default
action does the same than the "yes" action, if the user presses
"Yes", two install processes will run concurrently, conflicting
between them and resulting in an installation error shown (due
to one of the processes being unable to install the themes),
while, at the same time, the themes are correctly installed (by
the second process).

Also, if the user presses "No", the "no" action is just ignored
and nothing is done; but the next "default" action is honored
and the themes are installed; exactly the opposite of what the
user wanted.

This workaround ensures that only one action is honored for
each notification, and any extra action is ignored until the
install notification is closed and a new notification is
shown.

Fix https://github.com/canonical/snapd-desktop-integration/issues/52