Use systemd target status
Using gtk_init_check() seems to prematurely launch the portal
service, which is a problem. Also, the GDM session check is
quite fragile.

For all these reasons, this PR changes the way we check if we
are in a graphical session: now we ask systemd the current
status of the graphical-session.target unit, and we proceed to
launch the service only when it is active.