Add check based on DBus well-known names
Since, in some cases, snapd-desktop-integration is launched too
soon, and that triggers launching xdg-desktop-portal too soon
too (because several services from mutter aren't still available
in DBus, so xdg-desktop-portal presumes that they don't exist).

To really fix this, a change in snapd is required that allows to
specify that a service (snapd-desktop-integration in this case)
must be bound to the graphical-session target instead of the
default target.

But since this will require time, in the meantime this patch
should fix this problem: it waits until this well-known name does
appear in the bus:

    org.freedesktop.portal.Desktop

When that happens, we can assume that the desktop is ready, and
launch snapd-desktop-integration. It's an ugly hack, but it
should work until full graphical-session target support is added
to snapd.

Anyway, since it is not guaranteed that it is launched in any
desktop (has been tested only in Gnome Shell), a 10-seconds
timeout is added too, to ensure that it is launched safely.

Of course, this works along the previous method of asking for the
session type.

Fix https://bugs.launchpad.net/xdg-desktop-portal/+bug/2042301/comments/22