Fix test
The test test_refresh_monitor failed to build because NULL was
being passed to the constructor, which has no parameters. This
happened when building in 25.10, but not in older systems,
because in them, GCC presumes that a function declared with ()
has an indeterminated number of parameters.

This has been fixed, and also "void" has been added to all the
functions without parameters, to ensure that a compile error is
always triggered if a parameters is passed to these functions.